+++ components/sftp/utils.ts
@@ -244,40 +244,15 @@ export const getFileIcon = (entry: SftpFileEntry): React.ReactElement => {
     return React.createElement(FileCode, { size: 14 });
 };
 
-// Sort configuration types
-export type SortField = 'name' | 'size' | 'modified' | 'type';
-export type SortOrder = 'asc' | 'desc';
-
-// Column widths type
-export interface ColumnWidths {
-    name: number;
-    modified: number;
-    size: number;
-    type: number;
-}
-
-export type SftpColumnVisibility = Record<keyof ColumnWidths, boolean>;
-
-export const DEFAULT_SFTP_COLUMN_VISIBILITY: SftpColumnVisibility = {
-    name: true,
-    modified: true,
-    size: true,
-    type: true,
-};
-
-export const normalizeSftpColumnVisibility = (value: unknown): SftpColumnVisibility => {
-    if (!value || typeof value !== 'object' || Array.isArray(value)) {
-        return DEFAULT_SFTP_COLUMN_VISIBILITY;
-    }
-
-    const stored = value as Partial<Record<keyof ColumnWidths, unknown>>;
-    return {
-        name: true,
-        modified: stored.modified !== false,
-        size: stored.size !== false,
-        type: stored.type !== false,
-    };
-};
+// Sort / column layout — owned by application/state; re-exported for UI callers.
+export {
+  DEFAULT_SFTP_COLUMN_VISIBILITY,
+  normalizeSftpColumnVisibility,
+  type ColumnWidths,
+  type SftpColumnVisibility,
+  type SortField,
+  type SortOrder,
+} from '../../application/state/sftp/columnLayout';
 
 export const isSftpColumnMenuKey = (key: string, shiftKey: boolean): boolean =>
     key === 'ContextMenu' || (key === 'F10' && shiftKey);
